requirements.txt
django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pymysql==1.1.2
python-dotenv==1.0.1
django-cors-headers==4.4.0